requirements_cloud.txt
Flask==2.3.3
openai==1.3.0
numpy==1.24.3
scikit-learn==1.3.0
gunicorn==21.2.0
python-dotenv==1.0.0